+++ dtool/arctool/__init__.py
@@ -13,6 +13,16 @@ from dtool.utils import write_templated_file
 
 HERE = os.path.dirname(__file__)
 TEMPLATE_DIR = os.path.join(HERE, '..', 'templates')
+README_SCHEMA = [
+    ("project_name", u"project_name"),
+    ("dataset_name", u"dataset_name"),
+    ("confidential", False),
+    ("personally_identifiable_information", False),
+    ("owner_name", u"Your Name"),
+    ("owner_email", u"your.email@example.com"),
+    ("owner_username", u"namey"),
+    ("date", u"today"),
+]
 
 
 def new_archive_dataset(staging_path, descriptive_metadata):