+++ dtool/arctool/cli.py
@@ -13,9 +13,9 @@ from dtool import (
     Project,
     DtoolTypeError,
     NotDtoolObject,
-    DescriptiveMetadata,
 )
 from dtool.arctool import (
+    README_SCHEMA,
     readme_yml_is_valid,
     new_archive_dataset,
 )
@@ -26,8 +26,7 @@ from dtool.archive import (
     compress_archive,
 )
 from dtool.slurm import generate_slurm_script
-from dtool.utils import auto_metadata
-from dtool.clickutils import create_project
+from dtool.clickutils import create_project, generate_descriptive_metadata
 
 from fluent import sender
 
@@ -98,20 +97,7 @@ def cli_new_dataset(staging_path, project_metadata=dict()):
 
     logger.emit('pre_new_archive', {'staging_path': staging_path})
 
-    readme_info = [
-        ("project_name", u"project_name"),
-        ("dataset_name", u"dataset_name"),
-        ("confidential", False),
-        ("personally_identifiable_information", False),
-        ("owner_name", u"Your Name"),
-        ("owner_email", u"your.email@example.com"),
-        ("owner_username", u"namey"),
-        ("date", u"today"),
-    ]
-    descriptive_metadata = DescriptiveMetadata(readme_info)
-    descriptive_metadata.update(auto_metadata("nbi.ac.uk"))
-    descriptive_metadata.update(project_metadata)
-    descriptive_metadata.prompt_for_values()
+    descriptive_metadata = generate_descriptive_metadata(README_SCHEMA, '..')
 
     dataset, dataset_path, readme_path = new_archive_dataset(
         staging_path, descriptive_metadata)