+++ dtool/datatool/cli.py
@@ -11,12 +11,9 @@ from fluent import sender
 from dtool import (
     __version__,
     DataSet,
-    DescriptiveMetadata,
-    NotDtoolObject,
-    Collection,
 )
-from dtool.utils import auto_metadata
-from dtool.clickutils import create_project
+from dtool.clickutils import create_project, generate_descriptive_metadata
+from dtool.datatool import README_SCHEMA
 
 logger = sender.FluentSender('arctool', host='v0679', port=24224)
 
@@ -32,6 +29,20 @@ def cli(fluentd_host):
     logger.emit('cli_command', message)
 
 
+@cli.command()
+def markup():
+    descriptive_metadata = generate_descriptive_metadata(
+        README_SCHEMA, '..')
+
+    dataset_name = descriptive_metadata["dataset_name"]
+
+    descriptive_metadata.persist_to_path(
+        '.', template='datatool_dataset_README.yml')
+
+    ds = DataSet(dataset_name)
+    ds.persist_to_path('.')
+
+
 @cli.group()
 def new():
     pass
@@ -39,38 +50,20 @@ def new():
 
 @new.command()
 def dataset():
-    try:
-        collection = Collection.from_path('.')
-        parent_descriptive_metadata = collection.descriptive_metadata
-    except NotDtoolObject:
-        parent_descriptive_metadata = {}
 
-    readme_info = [
-        ("project_name", "project_name"),
-        ("dataset_name", "dataset_name"),
-        ("confidential", False),
-        ("personally_identifiable_information", False),
-        ("owner_name", "Your Name"),
-        ("owner_email", "your.email@example.com"),
-        ("owner_username", "namey"),
-        ("date", "today"),
-    ]
-    descriptive_metadata = DescriptiveMetadata(readme_info)
-    descriptive_metadata.update(auto_metadata("nbi.ac.uk"))
-    descriptive_metadata.update(parent_descriptive_metadata)
-    descriptive_metadata.prompt_for_values()
+    descriptive_metadata = generate_descriptive_metadata(
+        README_SCHEMA, '.')
+
     dataset_name = descriptive_metadata["dataset_name"]
-
     if os.path.isdir(dataset_name):
         raise OSError('Directory already exists: {}'.format(dataset_name))
-
     os.mkdir(dataset_name)
-
-    ds = DataSet(dataset_name, 'data')
-    ds.persist_to_path(dataset_name)
 
     descriptive_metadata.persist_to_path(
         dataset_name, template='datatool_dataset_README.yml')
+
+    ds = DataSet(dataset_name, 'data')
+    ds.persist_to_path(dataset_name)
 
 
 @new.command()