+++ CHANGELOG.rst
@@ -11,9 +11,13 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
+- ``datatool markup`` CLI command
+- ``dtool.metadata_from_path`` helper function
+
 Changed
 ^^^^^^^
 
+- datatool and arctool CLIs now use utility functions for new dataset and markup
 
 Deprecated
 ^^^^^^^^^^